backend/requirements.txt
langchain>=0.1.14
langchain-core>=0.1.14
langchain-community>=0.0.19
openai>=1.29.0
pinecone-client>=3.0.0
pydantic>=2.1.1
python-dotenv>=1.0.1
langsmith>=0.1.21
tqdm>=4.66.0
pypdf>=3.17.4
fastapi
uvicorn
python-dotenv
#add pnpm for frontend